src/app/requirements.txt
# Function dependencies
azure-functions>=1.15.0,<2.0.0
azure-identity>=1.13.0,<2.0.0
azure-monitor-opentelemetry>=1.0.0,<2.0.0
opentelemetry-instrumentation-requests>=0.40b0,<1.0.0
requests>=2.31.0,<3.0.0
